use dockerfile for native-ubuntu builds, use old approach only for wsl
and update docs and docs formatting

fix dockerfile
