+++ coinlib/README.md
@@ -85,13 +85,22 @@ place the library under a `build` directory.
 ## Building for Windows
 ### Building for Windows on Linux
 
-Cross-compile a secp256k1 DLL for Windows on an Ubuntu 20.04 host with `dart run coinlib:build_windows` or `dart run bin/build_windows.dart`.
+Cross-compile a secp256k1 DLL for Windows on an Ubuntu 20.04 host with
+`dart run coinlib:build_windows` or `dart run bin/build_windows.dart`.
 
 ### Building for Windows on Windows
 
-The above ["Building for Windows on Linux"](#building-for-windows-on-linux) can be completed in an Ubuntu 20.04 WSL2 instance on a Windows host.  You can also complete the build without installing Flutter to the WSL2 host by following [bitcoin-core/secp256k1's "Cross compiling" guide](https://github.com/bitcoin-core/secp256k1?tab=readme-ov-file#cross-compiling).
+Builds on Windows can be accomplished using WSL (Windows Subsystem for Linux).
+Cross-compile a secp256k1 DLL for Windows on an Ubuntu 20.04 WSL2 instance on a
+Windows host with `dart run coinlib:build_wsl` or `dart run bin/build_wsl.dart`,
+or complete the above ["Building for Windows on Linux"](#building-for-windows-on-linux)
+after installing Docker or Podman in WSL.  The build can also be completed 
+without installing Flutter to WSL by following 
+[bitcoin-core/secp256k1's "Cross compiling" guide](https://github.com/bitcoin-core/secp256k1?tab=readme-ov-file#cross-compiling).
 
-Building natively on Windows without WSL2 may also be possible by following [bitcoin-core/secp256k1's guide for Building on Windows](https://github.com/bitcoin-core/secp256k1?tab=readme-ov-file#building-on-windows) and placing the DLL at `src/secp256k1/lib/src/libsecp256k1-2.dll`.
+Building natively on Windows without WSL2 may also be possible by following 
+[bitcoin-core/secp256k1's guide for Building on Windows](https://github.com/bitcoin-core/secp256k1?tab=readme-ov-file#building-on-windows)
+and placing the DLL at `src/secp256k1/lib/src/libsecp256k1-2.dll`.
 
 ## Development
